+++ src/Event/Resource/AbstractResourceEvent.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ecourty\McpServerBundle\Event\Resource;
+
+abstract class AbstractResourceEvent
+{
+    public function __construct(
+        private readonly string $uri,
+    ) {
+    }
+
+    public function getUri(): string
+    {
+        return $this->uri;
+    }
+}