+++ src/Event/Resource/ResourceReadEvent.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ecourty\McpServerBundle\Event\Resource;
+
+class ResourceReadEvent extends AbstractResourceEvent
+{
+}