+++ src/Resource/TemplateResourceDefinition.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ecourty\McpServerBundle\Resource;
+
+/**
+ * Represents a templated resource definition (dynamic path).
+ *
+ * @see https://datatracker.ietf.org/doc/html/rfc6570
+ *
+ * @internal
+ */
+class TemplateResourceDefinition extends AbstractResourceDefinition
+{
+}