+++ src/Resources/config/services.php
@@ -24,6 +24,7 @@ return static function (ContainerConfigurator $container) {
             __DIR__ . '/../../Exception',
             __DIR__ . '/../../HttpFoundation',
             __DIR__ . '/../../Prompt',
+            __DIR__ . '/../../Resource',
             __DIR__ . '/../../Resources',
             __DIR__ . '/../../Tool',
             __DIR__ . '/../../McpServerBundle.php',